zebra: add proper netlink support for multihop IPv6 routes
On Linux, multihop IPv6 routes are implemented in a different
way. Instead of sending one route with additional hops as attributes,
we need to send each route to the kernel. Except the first one, those
routes should be appended.

With IPv4, the appropriate "ip route" command to add a multihop route
is:

    ip route add 10.0.1.0/24 proto quagga \
       nexthop via 10.234.95.10 dev eth0 \
       nexthop via 10.234.95.11 dev eth1

With IPv6, this is:

    ip -6 route add    2001:db8:97::/64 proto quagga via 2001:db8:1::2 dev eth0
    ip -6 route append 2001:db8:97::/64 proto quagga via 2001:db8:2::2 dev eth1

See the appropriate patch for the kernel:
 http://patchwork.ozlabs.org/patch/188562/

This patch hijacks the "single hop" case. When a new hop is about to
be added, the previous hop is sent to the kernel.