zclient: clarify when we can use nexthop and ifindex fields in zapi_ipvX
With IPv4, zapi_ipv4_route() is only called with either nexthops or
ifindexes. Never both of them. We enforce this with a warning and a
comment.

On the other hand, with IPv4, zapi_ipv6_route() is always called with
the same number of nexthops and ifindexes. Likewise, we enforce it
with a warning and a comment.