zebra: use ZEBRA_NEXTHOP_IPV6_IFINDEX to send interface-indexed IPv6 route
zapi_ipv6_route() was sending IPv6 routes by sending all the
gateways (as ZEBRA_NEXTHOP_IPV6) and all the ifindexes (as
ZEBRA_NEXTHOP_IFINDEX). On the other side, zebra was receiving the
routes and keep the last gateway and the last ifindex.

Because the number of gateways and the number of ifindexes is the
same, this works as expected (for a single hop route, for a multi hop
route, only the last route is kept).

To prepare the support for multihop route, we use
ZEBRA_NEXTHOP_IPV6_IFINDEX to send IPv6 routes which have both a
gateway and an interface (currently, this means all routes). The
receiving code is still awkward and will be fixed in the next commit.