ospf6d: merge ECMP routes when sending them to zebra
When several routes are identical (with respect to
`ospf6_route_cmp()`) and one of them is the best path, we consider all
of them to be the best path. When sending one of them to zebra, we
send them all as an ECMP route.

They will still appear as several routes in `show ipv6 ospf6 route`,
but they will be merged in `show ipv6 route`. `ospf6_route_is_best()`
will return true to any of them. Previous and next routes need to be
considered to build a multipath route.