EVPN: Exchange advertise-vni between bgpd and zebra
In order to get BGP to advertise VNIs and local MACs corresponding to
those VNIs, the operator has to configure 'advertise-vni'. This patch
implements the exchange of that information between bgpd and zebra and
actions in zebra upon receipt.

Note that this configuration controls only the advertisement of local
VNIs (and MACs) to peers which have negotiated the EVPN address-family.
Received VNIs will be accepted and processed, however they won't get
installed in zebra unless there is a corresponding local VNI (matching
based on BGP route target).

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-12034
Reviewed By: CCR-5021
Testing Done: Will be tested with a subsequent patch