bgpd: Fix code to deal with two-level routing table
The routing table for certain address families - L3VPN, ENCAP and EVPN -
is organized as a two-level table with Route Distinguisher (RD) at the
top level pointing to a second-level table of prefixes. The code does
not handle this well in some cases and this patch fixes some such
area.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-12261
Reviewed By: CCR-5052
Testing Done: Manual