zebra: Ensure VxLAN interface status change is handled
Ensure zebra reacts properly to change in VxLAN interface status. While VxLAN
interfaces are learnt and hash table maintained in zebra irrespective of
status, the local VNIs are informed to BGP only when the underlying interface
is up and removed from BGP when not up. Likewise, remote VTEPs are installed
only if the interface is up.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-12294
Reviewed By: CCR-5077
Testing Done: Manual