bgpd: Remove 'auto' option for RD and RT configuration for EVPN
Since the default is to automatically derive the RD and RT automatically
for EVPN (using RouterId:VNI and AS:VNI respectively), it is not useful or
necessary to also have an 'auto' configuration option. This commit removes
the 'auto' option from the configuration.

When the operator configures an RD or RT, that will be used in place of the
automatically derived value. Additional RTs can be added for "import", but
further configuration of RD or RT for "export" will be treated as a "replace".
If the RD to RT is removed, the corresponding parameter will default back
to being automatically derived.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>