bgpd: VNI configuration updates
Cleanup code and implement some behavior changes with respect to VNI
configuration in BGP:

1. Define and use flags for configuration being set rather than have
flags for the default configuration.

2. Cleanup how a VNI is determined as "configured" - it is based on any
VNI parameter (RD, Import RT or Export RT) being configured. If the
operator just specifies "vni <value>" to "create" a VNI, that is not
treated as a "configured" VNI i.e., while it will show up in the output
of "show bgp evpn vni", it will not show up in the saved configuration.

3. If a configured VNI is attempted to be deleted, it will be permitted
only if the VNI is not defined in the kernel; otherwise the operator has
to first disable/delete it in the kernel. This is to simplify internal
operation, as in this case, the VNI will reincarnate with default RD and
RT, which will require all routes to be withdrawn and re-advertised.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>