zebra: Handle deleted interfaces properly
When an interface is deleted, zebra does not actually delete the
interface data structure but keeps it with the ifindex reset. The
associated zebra interface structure is also retained, but some other
information is purged. This is due to legacy reasons. Due to this
behavior, a "stale" interface is wrongly considered to be a bridge
slave resulting in an exception.

The fix is to ensure that information related to an interface being a
bridge slave is reset when the interface is deleted. Also, deleted
interfaces are skipped when trying to map a bridge to its slaves.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-13537
Reviewed By: None
Testing Done: Manual (recreate problem, verify fix)