bgpd: Handle VxLAN tunnel IP change
When the VxLAN tunnel local IP changes, MACIP routes need to be
re-advertised with the new tunnel IP as the next hop. The multicast
route also needs to be updated, but that actually requires the route
to be withdrawn and re-advertised since the local tunnel IP is part of
the key.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-13552
Reviewed By:
Testing Done: Manual