zebra: Use EVPN enable/disable to control VNI and MAC learning
Expand the use of EVPN enable/disable (the command is 'advertise-vni') to
control the building of the VNI hash table and the processing of MAC table
notifications by zebra. This is done to ensure that zebra doesn't bother
about VNIs and MACs when EVPN is not in use.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-13657
Reviewed By:
Testing Done: Manual