redhat: fix up spec.in to build in a mock chroot
Review: CCR-5413
Testing-done: built in an epel7 mock chroot and on a centos 7 VM

The check to set the %initsystem macro failed for systemd-based
systems inside of a chroot, because `systemctl' has different behavior
while chrooted.  Fix it up to check the symlink target of /sbin/init
instead, using bash's '-ef' test flag.

Also fixed a typo (ospfd6d -> ospf6d) in the same file which was
causing the build to fail.

Signed-off-by: Silas McCroskey <smccroskey@cumulusnetworks.com>
Reviewed-by: Donald Sharp <sharpd@cumulusnetworks.com>