bgpd: Remove VNI from prefix structure
The VNI is not part of the "route" itself but is an associated
parameter that should be exchanged with the route (in the field
originally defined for the MPLS label). Remove VNI from the prefix
structure definition.

Note: A subsequent commit will address passing the VNI correctly.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14117
Reviewed By: None
Testing Done: Manual