bgpd: Carry VNI in label field of EVPN type-2 route
Ensure that the VNI is carried properly in the label field of the EVPN
type-2 route (as per RFC 7432 and draft-ietf-bess-evpn-overlay). On
the receive side, the VNI is stored in the BGP route. Right now, the
VNI is not passed down to zebra because we only support global VNIs,
which means that the VNI into which the received route is imported
(based on its RT attributes) should be the same as the VNI value in
the received route. This part needs to be changed when we support local
VNIs.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14118
Reviewed By: None
Testing Done: Manual