bgpd: Ensure that VNI configuration is ordered correctly
Currently, an RD cannot be configured for a VNI without first disabling
it. When Quagga comes up, reading its configuration and interaction
between the daemons and with the kernel are interspersed. Unless any
RD configuration is processed by BGP before VNIs are learnt from the
kernel, the configuration will be rejected/lost. Ensure this is done
by ordering the configuration correctly.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14358
Reviewed By: None
Testing Done: Manual

Note: The restriction on RD configuration will be removed in a future
release. Also, this change was available earlier but missed in commit.