zebra: Remove invalid asserts
While normally, a layer-2 interface that is enslaved to a bridge should
have its "L2 information", it may be missing in some cases - e.g., when
some interfaces have been deleted (due to the way zebra handles the delete
of an interface). Ensure that the incorrect assertions are removed.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14444
Reviewed By: None
Testing Done: Manual