bgpd: Handle MAC Mobility extended community for EVPN routes
Implement handling of MAC Mobility extended community as per RFC 7432.
Note that this change only implements the following:
a) Record presence of MAC Mobility extended community in a route's
attribute; if not present, it is recorded as 0.
b) When a local MAC is detected for which a remote entry (or entries)
already exist, set the MAC Mobility sequence number for the local route
appropriately
c) In route selection, for EVPN routes, factor in the MAC Mobility
sequence number prior to using route type (in order to allow a "more
recent" remote route to win over a local route).
d) When a remote MAC becomes best (due to MAC mobility) and the prior
best route was a local MAC, delete and withdraw the local MAC.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14114
Reviewed By: None
Testing Done: Manual