zebra: Read MAC FDB upon VxLAN interface creation, other refinements
Quagga maintains its MAC database against the VNI, so after a VxLAN
interface is created, it has to read the MAC FDB corresponding to
the VxLAN interface's bridge (and associated VLAN, if any) in order
to learn about existing local MACs. This also requires refinements in
handling change to bridge membership for a VxLAN interface and change
to its access VLAN.

A consequence of this change is that the VNI will be informed to BGP
only when the VxLAN interface is mapped to a bridge and is up.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14440
Reviewed By: None
Testing Done: Manual