zebra: Handle change for bridge interface
Whether a bridge is VLAN-aware or not may be notified subsequent to the
bridge interface creation. Zebra was not handling this well resulting
in the bridge being mistaken to be non-VLAN aware after some operations,
leading to misprogramming of MAC entries and other misbehavior. Ensure
that the change handling is correct in zebra.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14674
Reviewed By: None
Testing Done: Manual