bgpd: Allow dynamic change to RD and handle it
The change of RD - either a configured value or deleting the existing
configured value - was not allowed on a VNI if the VNI was defined in
the system (kernel). Implement changes to allow it. This involves
deleting and withdrawing routes with the prior RD and then advertising
routes with the new RD.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14352
Reviewed By: None
Testing Done: Manual