bgpd: Allow delete of VNI that exists in the system
Allow a VNI (configuration) to be deleted even if it exists in the
system (kernel). The result would be that the VNI continues to exist
and is advertised by BGP along with all local MACs corresponding to
it, except it will be using automatically derived RDs and RTs. This
operation will result in EVPN routes being withdrawn and re-advertised
to peers.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14521
Reviewed By: None
Testing Done: Manual