bgpd: Fix partial match of route target for import
To facilitate EVPN route exchange between EBGP peers without needing
additional configuration, a partial match of the received route targets
(in the route) is supported, as long as the import route-target for a
VNI is auto-derived and not configured. Ensure that this partial match
correctly acts as *:VNI, ignoring all other fields of the received route
target other than the local-admin field, which is expected to contain
the VNI value.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14759
Reviewed By: None
Testing Done: Manual