bgpd: Extract and store MAC mobility extended community correctly
Besides saving the MAC mobility extended community itself, the value of the
MAC mobility sequence number is extracted and stored separately for ease of
access (otherwise, all extended communities need to be searched). This
extracted value is not being updated/set in some situations resulting in
misbehavior. Ensure that the value is extracted and set in all scenarios.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-14792
Reviewed By: None
Testing Done: Manual